+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Sun - 4/25/21 | Sprint Grape 🍇](#)
+- ✨ Impl populating the sidebar ('GTx ISYE6501' menu items) from Firestore
 - ✨ Add '/scripts' and 'firebase-admin' in order to easily populate Firestore
 - ♻️ Rename `Secure.tsx` to `FileViewer.tsx`. Properly implement `FileViewer.tsx`
 - 💄 Beautify the 'Secure Asset' page to show a nice message if user is unauthorized to view asset